+++ linker.ld
@@ -232,7 +232,7 @@ SECTIONS {
         src/code_0800ECAC.o(.text);
         asm/sub_0800F044.o(.text);
         src/code_0800ECAC_2.o(.text);
-        src/code_08019CB8.o(.text);
+        src/treasures.o(.text);
         src/multi_08019F28.o(.text);
         asm/goal_game_bonus.o(.text);
         src/shard.o(.text);
