+++ src/code_0801C6F8.c
@@ -3,6 +3,7 @@
 #include "functions.h"
 #include "palette.h"
 #include "save.h"
+#include "treasures.h"
 
 static void sub_0801C79C(void);
 static void nullsub_107(struct Task *);
@@ -322,7 +323,7 @@ static void sub_0801CE74(struct Unk_0801C6F8 *var)
 {
     struct Unk_02022930_0 *unkStruct;
 
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
         m4aSongNumStart(0x213);
     else
         m4aSongNumStart(0x212);
@@ -330,7 +331,7 @@ static void sub_0801CE74(struct Unk_0801C6F8 *var)
     unkStruct->unk8 |= 0x180;
     unkStruct->unk6 = 0;
     unkStruct->unk4 = 6;
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
         unkStruct->unk4 |= 1;
     var->unk20C &= ~0x200;
     var->unk20C |= 0x20000000;
@@ -340,7 +341,7 @@ static void sub_0801CE74(struct Unk_0801C6F8 *var)
 
 static void sub_0801CF00(struct Unk_0801C6F8 *var)
 {
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
     {
         var->unk20C &= ~0xFF;
         var->unk20C |= 0x100;
@@ -364,9 +365,9 @@ static void sub_0801CF78(struct Unk_0801C6F8 *var)
     unkStruct->unkA = unkStruct->unkA >> 2;
     unkStruct->unk6 = 0;
     unkStruct->unk4 = 6;
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
         unkStruct->unk4 |= 1;
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
         SpriteSomething(&sprite, 0x6000000, 0x2D3, 1, 0xFF, 0, 0, 0, 0, 0x10, 0, 0x80000);
     else
         SpriteSomething(&sprite, 0x6000000, 0x2D3, 0, 0xFF, 0, 0, 0, 0, 0x10, 0, 0x80000);
@@ -377,7 +378,7 @@ static void sub_0801CF78(struct Unk_0801C6F8 *var)
 
 static void sub_0801D080(struct Unk_0801C6F8 *var)
 {
-    if (numShards() > 6)
+    if (numShardsCollected() > 6)
     {
         u16 i;
 
@@ -642,7 +643,7 @@ static void sub_0801D7B0(struct Unk_0801C6F8 *var)
 {
     if (var->unk21C++ > 0x28)
     {
-        if (numShards() > 6)
+        if (numShardsCollected() > 6)
             var->unk208 = sub_0801D870;
         else
             var->unk208 = sub_0801D7F8;
